+++ domain/src/main/kotlin/omok/Application.kt
@@ -1,5 +1,8 @@
 package omok
 
+import omok.controller.Controller
+import omok.view.GameView
+
 fun main() {
     runCatching {
         Controller(GameView(), OmokGame()).gameStart()
